travis: update to build docs and a bigger matrix
Test against more versions of Rust with each push and ensure that
Rustache stays compatible with Rust 1.3.0, current stable, current beta
and detect potential breakage with the nightly but not fail. Rust 1.3.0
is selected due to that being the minimum version of rustache's
dependencies.  Additionally this keeps the docs up to date by building
them with each push. This change is based on the blog post at:
https://pascalhertleif.de/artikel/good-practices-for-writing-rust-libraries/

Note to maintainers: You must set a value for env.global.secure based on
the 'Automatically Render Documentation' section of the site above.