drop the conversion from Path to HashBuilder
This API is likely confusing whether the user is expected to supply the
path of the data set or the template. Additionally data sets are allowed
to be arrays or objects and in this case this API only supported files
that were objects. Its not mentioned heavily so remove it now before it
gains much use.